feat: updated keyple to 0.9-alpha2 (Windows)

+++ component/keyple-calypso/src/main/transaction/PoSecuritySettings.cpp
@@ -35,7 +35,7 @@ const RatificationMode PoSecuritySettings::mDefaultRatificationMode =
 const PinTransmissionMode PoSecuritySettings::mDefaultPinTransmissionMode =
     PinTransmissionMode::ENCRYPTED;
 const std::shared_ptr<KeyReference> PoSecuritySettings::mDefaultPinCipheringKey =
-    std::make_shared<KeyReference>(0, 0);
+    std::make_shared<KeyReference>(static_cast<uint8_t>(0), static_cast<uint8_t>(0));
 const LogRead PoSecuritySettings::mDefaultSvGetLogReadMode = LogRead::SINGLE;
 const NegativeBalance PoSecuritySettings::mDefaultSvNegativeBalance = NegativeBalance::FORBIDDEN;
